+++ tests/unit/shared/repository/GithubRepositoryTest.php
@@ -1,7 +1,6 @@
 <?php
 namespace PharIo\Phive;
 
-use PharIo\Version\Version;
 use PHPUnit\Framework\TestCase;
 
 /**
@@ -17,18 +16,19 @@ class GithubRepositoryTest extends TestCase {
         $requestedPhar->method('getAlias')->willReturn($pharAlias);
 
         $entry1 = $this->getGithubEntry('5.3.0', 'https://example.com/foo-5.3.0.phar');
-        $entry2 = $this->getGithubEntry('5.2.11', 'https://example.com/broken');
-        $entry3 = $this->getGithubEntry('5.2.12', 'https://example.com/foo-5.2.12.phar');
+        $entry2 = $this->getGithubEntry('foo-5.3.0', 'https://example.com/foo-5.3.0.phar');
+        $entry3 = $this->getGithubEntry('5.2.11', 'https://example.com/broken');
+        $entry4 = $this->getGithubEntry('5.2.12', 'https://example.com/foo-5.2.12.phar');
 
         $jsonData = $this->getJsonDataMock();
         $jsonData->method('getParsed')
-            ->willReturn([$entry1, $entry2, $entry3]);
+            ->willReturn([$entry1, $entry2, $entry3, $entry4]);
 
         $expectedReleases = new ReleaseCollection();
         $expectedReleases->add(
             new Release(
                 'foo',
-                new Version('5.3.0'),
+                new GitHubVersion('5.3.0'),
                 new PharUrl('https://example.com/foo-5.3.0.phar'),
                 new Url('https://example.com/foo-5.3.0.phar.asc')
             )
@@ -36,7 +36,7 @@ class GithubRepositoryTest extends TestCase {
         $expectedReleases->add(
             new Release(
                 'foo',
-                new Version('5.2.12'),
+                new GitHubVersion('5.2.12'),
                 new PharUrl('https://example.com/foo-5.2.12.phar'),
                 new Url('https://example.com/foo-5.2.12.phar.asc')
             )